Migrate select and password prompts
