+++ inquire/src/prompts/dateselect/actions.rs
@@ -19,10 +19,10 @@ impl InnerAction<DateSelectConfig> for DateSelectPromptAction {
     fn from_key(key: Key, config: &DateSelectConfig) -> Option<Self> {
         if config.vim_mode {
             let action = match key {
-                Key::Char('k', KeyModifiers::NONE) => Some(DateSelectPromptAction::GoToPrevWeek),
-                Key::Char('j', KeyModifiers::NONE) => Some(DateSelectPromptAction::GoToNextWeek),
-                Key::Char('h', KeyModifiers::NONE) => Some(DateSelectPromptAction::GoToPrevDay),
-                Key::Char('l', KeyModifiers::NONE) => Some(DateSelectPromptAction::GoToNextDay),
+                Key::Char('k', KeyModifiers::NONE) => Some(Self::GoToPrevWeek),
+                Key::Char('j', KeyModifiers::NONE) => Some(Self::GoToNextWeek),
+                Key::Char('h', KeyModifiers::NONE) => Some(Self::GoToPrevDay),
+                Key::Char('l', KeyModifiers::NONE) => Some(Self::GoToNextDay),
                 _ => None,
             };
 
@@ -31,16 +31,18 @@ impl InnerAction<DateSelectConfig> for DateSelectPromptAction {
             }
         }
 
-        match key {
-            Key::Left(KeyModifiers::NONE) => Some(DateSelectPromptAction::GoToPrevDay),
-            Key::Right(KeyModifiers::NONE) => Some(DateSelectPromptAction::GoToNextDay),
-            Key::Up(KeyModifiers::NONE) => Some(DateSelectPromptAction::GoToPrevWeek),
-            Key::Down(KeyModifiers::NONE) | Key::Tab => Some(DateSelectPromptAction::GoToNextWeek),
-            Key::Left(KeyModifiers::CONTROL) => Some(DateSelectPromptAction::GoToPrevMonth),
-            Key::Right(KeyModifiers::CONTROL) => Some(DateSelectPromptAction::GoToNextMonth),
-            Key::Up(KeyModifiers::CONTROL) => Some(DateSelectPromptAction::GoToPrevYear),
-            Key::Down(KeyModifiers::CONTROL) => Some(DateSelectPromptAction::GoToNextYear),
-            _ => None,
-        }
+        let action = match key {
+            Key::Left(KeyModifiers::NONE) => Self::GoToPrevDay,
+            Key::Right(KeyModifiers::NONE) => Self::GoToNextDay,
+            Key::Up(KeyModifiers::NONE) => Self::GoToPrevWeek,
+            Key::Down(KeyModifiers::NONE) | Key::Tab => Self::GoToNextWeek,
+            Key::Left(KeyModifiers::CONTROL) => Self::GoToPrevMonth,
+            Key::Right(KeyModifiers::CONTROL) => Self::GoToNextMonth,
+            Key::Up(KeyModifiers::CONTROL) => Self::GoToPrevYear,
+            Key::Down(KeyModifiers::CONTROL) => Self::GoToNextYear,
+            _ => return None,
+        };
+
+        Some(action)
     }
 }
